rename fabric_rma_van as fabric_van
